Widen abs scroll-timer temp to int in cMenuRun
int m = gCMenuScrollTimer drops the redundant s16 truncation extsh on
the (m<0?-m:m) abs idiom, matching target's lha;neg;cmpwi form (no
extsh). 97.42->98.13.

+++ src/main/dll/dll_0000_gameui.c
@@ -2079,7 +2079,7 @@ void cMenuRun(void)
             }
             if ((cy <= -0xa && gCMenuPrevStickY > -0xa) || cy < -0x3c)
             {
-                s16 m = gCMenuScrollTimer;
+                int m = gCMenuScrollTimer;
                 if (m < 0) m = -m;
                 if (m < 8 && gCMenuScrollLock == 0 && lbl_803DD79A == 0)
                 {
@@ -2093,7 +2093,7 @@ void cMenuRun(void)
             }
             if ((cy >= 0xa && gCMenuPrevStickY < 0xa) || cy > 0x3c)
             {
-                s16 m = gCMenuScrollTimer;
+                int m = gCMenuScrollTimer;
                 if (m < 0) m = -m;
                 if (m < 8 && gCMenuScrollLock == 0 && lbl_803DD79A == 0)
                 {
